calendar-summar/requirements.txt
google-api-python-client
google-auth-oauthlib
slack_sdk
